% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notion.R-package.R
\docType{package}
\name{notion.R-package}
\alias{notion.R}
\alias{notion.R-package}
\title{notion.R: Notion API for R}
\description{
Notion is a new tool that blends many different work tools into one. This package provides utilities for dealing with the Notion API from R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tjpalanca/notion.R}
  \item Report bugs at \url{https://github.com/tjpalanca/notion.R/issues}
}

}
\author{
\strong{Maintainer}: Troy James Palanca \email{mail@tjpalanca.com} (\href{https://orcid.org/0000-0002-0610-9213}{ORCID})

}
\keyword{internal}
