% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_api_call}
\alias{search_api_call}
\title{Notion Search API}
\usage{
search_api_call(
  query = "",
  sort_direction = c("ascending", "descending"),
  sort_timestamp = c("last_edited_time"),
  filter_value = c("page", "database"),
  filter_property = c("object"),
  start_cursor = NULL,
  page_size = 100L
)
}
\arguments{
\item{query}{\code{str} limits which pages are returned by comparing with title}
}
\description{
\href{https://developers.notion.com/reference/post-search}{Documentation}
}
