% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-list.R
\name{user_list}
\alias{user_list}
\title{List all users}
\usage{
user_list(n_max = Inf)
}
\arguments{
\item{n_max}{(int) maximum number of users to return}
}
\description{
\href{https://developers.notion.com/reference/get-users}{API Reference}
}
